lucky_kangaroo/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Mail==0.9.1
Flask-Caching==2.0.2
Flask-Limiter==3.5.0
python-dotenv==1.0.0
Pillow==10.0.1
rembg==2.0.38
stripe==7.0.0
openai==0.28.1
sentry-sdk[flask]==1.34.0
redis==5.0.1
python-dateutil==2.8.2
